refactor(newSpecial):공통 타입 정의, UI 로직과 데이터 분리, 기타 코드 정리
