+++ src/domains/newSpecial/components/blocks/GameSlider.tsx
@@ -2,72 +2,8 @@ import { Link } from 'react-router-dom'
 import { useState } from 'react'
 import { Swiper, SwiperSlide } from 'swiper/react'
 import type { Swiper as SwiperClass } from 'swiper'
-
 import { FreeMode, EffectFade, Navigation, Thumbs } from 'swiper/modules'
-
-//--------------- 이미지 ---------------//
-import gameSlideImg1 from '@/assets/imgs/temp/sub_gameSlider_img1.png'
-import gameSlideImg2 from '@/assets/imgs/temp/sub_gameSlider_img2.png'
-import gameSlideImg3 from '@/assets/imgs/temp/sub_gameSlider_img3.png'
-import gameSlideImg4 from '@/assets/imgs/temp/sub_gameSlider_img4.png'
-import gameSlideImg5 from '@/assets/imgs/temp/sub_gameSlider_img5.png'
-
-import gameSlideThumbImg1 from '@/assets/imgs/temp/sub_gameSlider_ThumbImg1.png'
-import gameSlideThumbImg2 from '@/assets/imgs/temp/sub_gameSlider_ThumbImg2.png'
-import gameSlideThumbImg3 from '@/assets/imgs/temp/sub_gameSlider_ThumbImg3.png'
-import gameSlideThumbImg4 from '@/assets/imgs/temp/sub_gameSlider_ThumbImg4.png'
-import gameSlideThumbImg5 from '@/assets/imgs/temp/sub_gameSlider_ThumbImg5.png'
-
-//--------------- 데이터 ---------------//
-const gameSlideData = [
-  {
-    name: 'THE FINALS',
-    description: `거칠고 치열한 자연의 습격. 시시각각 역동적으로 그 모습을 바꾸는
-              필드. 양면성을 지닌 세계를 살아가는 몬스터와 사람들의 이야기.
-              더욱더 발전한 헌팅 액션과 끊임없는 몰입감을 추구하는 궁극의 사냥
-              체험이 당신을 기다리고 있다.`,
-    img: gameSlideImg1,
-    thumb: gameSlideThumbImg1,
-  },
-  {
-    name: 'PUBG: BATTLEGROUNDS',
-    description: `PUBG: BATTLEGROUNDS를 무료로 플레이 하세요. 다양한 전장에서 전략적
-              위치를 선점하고 무기와 장비를 확보해 최후의 1인이 되기 위한 생존의
-              사투를 펼칩니다. 친구들과 함께 팀을 만들어 배틀로얄 건플레이
-              장르의 선구자인 PUBG: BATTLEGROUNDS만이 선사하는 긴장감 넘치는
-              경험을 위해 도전하세요.`,
-    img: gameSlideImg2,
-    thumb: gameSlideThumbImg2,
-  },
-  {
-    name: 'Sons Of The Forest',
-    description: `외딴 섬에서 실종된 억만장자를 찾기 위해 파견된 당신은 식인종으로
-              가득한 지옥을 마주하게 됩니다. 이 무시무시한 오픈 월드 기반의 공포
-              서바이벌 시뮬레이터에서 혼자 또는 친구들과 함께 무기를 제작하고
-              건물을 지으며 생존을 위해 고군분투하세요.`,
-    img: gameSlideImg3,
-    thumb: gameSlideThumbImg3,
-  },
-  {
-    name: 'THE FINALS',
-    description: `전 세계 명소를 배경으로 한 가상의 세계에서 펼쳐지는 전투 중심의
-              게임쇼, THE FINALS에 참여하세요! 적 팀의 취약점을 파악하고,
-              불가능해 보이지만 모두 부셔 버릴 수 있는 특별한 전장에서 팀원들과
-              싸워보십시오, 나만의 플레이 스타일로 토너먼트의 승리를 차지해
-              보세요.`,
-    img: gameSlideImg4,
-    thumb: gameSlideThumbImg4,
-  },
-  {
-    name: 'Ready or Not',
-    description: `로스수에뇨스 – LSPD는 로스수에뇨스 지역 전역에서 폭력 범죄가
-              급증하고 있다고 발표했습니다. 이에 인질, 폭탄 위협, 바리케이드를
-              세우고 있는 용의자 등 매우 위험한 상황에 대응하기 위해 SWAT 팀이
-              파견되었습니다. 시민 여러분은 집에 머무시기 바랍니다.`,
-    img: gameSlideImg5,
-    thumb: gameSlideThumbImg5,
-  },
-]
+import { gameSlidesData } from '@newSpecial/constants'
 
 export const GameSlider = () => {
   const [thumbsSwiper, setThumbsSwiper] = useState<SwiperClass | null>(null)
@@ -82,7 +18,7 @@ export const GameSlider = () => {
         speed={700}
         effect={'fade'}
       >
-        {gameSlideData.map((game, index) => (
+        {gameSlidesData.map((game, index) => (
           <SwiperSlide key={index}>
             <div className="game_img">
               <img src={game.img} alt={game.name} />
@@ -104,7 +40,7 @@ export const GameSlider = () => {
         modules={[FreeMode, Navigation, Thumbs]}
         className="thumb"
       >
-        {gameSlideData.map((game, index) => (
+        {gameSlidesData.map((game, index) => (
           <SwiperSlide key={index}>
             <div className="game_img">
               <img src={game.thumb} alt={game.name} />
