+++ src/domains/newSpecial/components/blocks/UpcomingGames.tsx
@@ -1,16 +1,5 @@
 import { Link } from 'react-router-dom'
-
-// --------------- 이미지 --------------- //
-import upcomingGamesImg1 from '@/assets/imgs/temp/sub_upcomingGames_img1.png'
-import upcomingGamesImg2 from '@/assets/imgs/temp/sub_upcomingGames_img2.png'
-import upcomingGamesImg3 from '@/assets/imgs/temp/sub_upcomingGames_img3.png'
-
-// 데이터 배열
-const upcomingGamesData = [
-  { id: 1, img: upcomingGamesImg1 },
-  { id: 2, img: upcomingGamesImg2 },
-  { id: 3, img: upcomingGamesImg3 },
-]
+import { upcomingGamesData } from '@newSpecial/constants'
 
 export const UpcomingGames = () => {
   return (
