+++ src/domains/newSpecial/components/index.ts
@@ -5,3 +5,6 @@ export * from './blocks/PopularNewGames'
 export * from './blocks/NewProductGames'
 export * from './blocks/UpcomingGames'
 export * from './blocks/UpcomingList'
+
+export * from './units/GameCard'
+export * from './units/BannerItem'
