+++ src/domains/newSpecial/constants/bannerData.ts
@@ -1,0 +1,19 @@
+import newGameBannerImg1 from '@/assets/imgs/temp/sub_newGame_bannerImg1.png'
+import newGameBannerImg2 from '@/assets/imgs/temp/sub_newGame_bannerImg2.png'
+
+export const bannerData = [
+  {
+    img: newGameBannerImg1,
+    name: '메타포: 리판타지오',
+    description:
+      '『메타포: 리판타지오』는 국왕 암살 후 돌연 시작된 선거에 휘말려 새로운 왕의 후보자가 된 소년의 운명을 그린 판타지 RPG입니다. 세계를 방랑하며 하루하루를 자유롭게 보내고, 다양한 클래스와 파티를 활용하여 강적과 맞서 싸우세요. "환상이기에 가능한 현실의 여행"이 시작됩니다.',
+    price: '₩79,800',
+  },
+  {
+    img: newGameBannerImg2,
+    name: 'ZERO Sievert',
+    description:
+      'ZERO Sievert는 알고리즘에 따라 생성된 황무지를 뒤지고 장비를 약탈하고 황폐해진 세계에 남겨진 것을 탐험하며 도전하는 강렬한 탑다운 익스트랙션 슈터 게임입니다. 여러분에게 불리한 상황이 닥치면, 생존 그 이상의 일을 해내야 할 것입니다…',
+    price: '₩21,500',
+  },
+]
