+++ src/domains/newSpecial/constants/gameCategories.ts
@@ -1,0 +1,10 @@
+export const gameCategories = [
+  '특집',
+  '액션',
+  '어드벤처',
+  '캐주얼',
+  'RPG',
+  '시뮬레이션',
+  '전략',
+  '컨트롤러',
+] as const
