+++ src/domains/newSpecial/constants/upcomingGamesData.ts
@@ -1,0 +1,10 @@
+import upcomingGamesImg1 from '@/assets/imgs/temp/sub_upcomingGames_img1.png'
+import upcomingGamesImg2 from '@/assets/imgs/temp/sub_upcomingGames_img2.png'
+import upcomingGamesImg3 from '@/assets/imgs/temp/sub_upcomingGames_img3.png'
+
+// 데이터 배열
+export const upcomingGamesData = [
+  { id: 1, img: upcomingGamesImg1 },
+  { id: 2, img: upcomingGamesImg2 },
+  { id: 3, img: upcomingGamesImg3 },
+]
